"ads1_m14_a1: 1/4 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ assignments/m14/14.1 Dynamic Median Finding/Solution.java
@@ -3,7 +3,7 @@ import java.lang.*;
 class Solution {
 	public static void main(String[] args) {
 		Scanner scan = new Scanner(System.in);
-		int number = scan.nextInt();
+		long number = scan.nextInt();
 		MinPQ<Double> min = new MinPQ<Double>();
 		MaxPQ<Double> max = new MaxPQ<Double>();
 		double median = 0.0;
